requirements.txt
aws-cdk-lib==2.82.0
aws-cdk.aws-lambda-python-alpha==2.82.0a0
constructs>=10.0.0,<11.0.0
